chore(fe): separate component file
